+++ ios/flutter_litert/Sources/flutter_litert/FlutterLitertPlugin.swift
@@ -1,8 +1,12 @@
 import Flutter
 import UIKit
+import flutter_litert_custom_ops
+import flutter_litert_delegate_symbols
 
 public class FlutterLitertPlugin: NSObject, FlutterPlugin {
   public static func register(with registrar: FlutterPluginRegistrar) {
+    FlutterLitertRetainFfiSymbols()
+    FlutterLitertRetainCustomOps()
     let channel = FlutterMethodChannel(name: "flutter_litert", binaryMessenger: registrar.messenger())
     let instance = FlutterLitertPlugin()
     registrar.addMethodCallDelegate(instance, channel: channel)